Further optimization, saves a further 6% of the interrupt execution time
Added a 3rd interrupt handler on TCA CMP1 and moving the call to playing->stop() there. This saves 6% (about 125 ns) from the execution time (not only when we call stop but every time the OVF interrupt runs), probably because the compiler can optimize better if no function is called in the handler.